chore: move routers folder to backend folder + change the api.py to follow development
