+++ api.py
@@ -24,7 +24,7 @@ from utils.general import validate_openai_api_key
 from utils.log_tool import set_color
 from utils.statistics import update_statistic_per_query
 
-from src.routers import models
+from backend.routers import models
 
 app = FastAPI()
 
@@ -33,8 +33,8 @@ app = FastAPI()
 # in your .env file depending on your frontend port, 8080 or 5173 in this case.
 
 CORS_ALLOW_ORIGIN = [
-    "http://localhost:8080",  # Open-WebUI production server
-    "http://localhost:5173",  # Open-WebUI development server
+    "http://localhost:8080",  # Frontend production server
+    "http://localhost:5173",  # Frontend development server
 ]
 
 # Allow CORS for the specified origins
